Increased type system around `WavePacket`
So far, this is a lot of code for just a little bit of convinience when 
creating coherent states in a flat mode.

However, this hides the `FlatMode()` implementation from the user, which 
is actually not a proper mode and also allows me later the write a 
version of `_generateLindbladian()`, which is optimised for continuous 
waves (time independant system).